Move framework stuff to folder Add icon for CircuitBreaker Use record copier to solve access exception
