--- AlteryxDateParser/ConnectionState.cs
@@ -1,9 +1,0 @@
-﻿namespace JDunkerley.Alteryx
-{
-    public enum ConnectionState
-    {
-        Added,
-        InitCalled,
-        Closed
-    }
-}
