+++ AlteryxDateParser/Framework/BaseEngine.cs
@@ -1,9 +1,8 @@
-namespace JDunkerley.Alteryx
+namespace JDunkerley.Alteryx.Framework
 {
     using System;
     using System.Collections.Generic;
     using System.Reflection;
-    using System.Threading;
     using System.Xml;
     using System.Xml.Serialization;
